# api/openapi.yaml
openapi: 3.1.0
info:
  title: OCS Testbench API
  version: 0.2.0
  description: |
    REST API for the OCS Testbench UI and external clients.

    This specification is the single source of truth for the HTTP contract
    between the Go backend (`internal/api`) and the React frontend (`web/`).
    Generated TypeScript types live at `web/src/api/schema.d.ts` — regenerate
    after any change to this file.

    See `docs/ARCHITECTURE.md` for behavioural semantics (state machines,
    invariants, variable resolution rules). This document is the wire
    contract; it does not restate the semantics.

    Real-time server → client events (peer status changes, execution
    progress, pause / resume transitions) are delivered via the SSE
    endpoint at `GET /events`.
servers:
  - url: /api/v1
    description: Current server (relative to origin)
...
paths:
...
  /scenarios:
    get:
      tags: [Scenarios]
      operationId: listScenarios
      summary: List scenarios (system starters + user scenarios)
      description: |
        Returns both system starters (`origin: "system"`, immutable) and
        user scenarios (`origin: "user"`). Client-side filtering expected;
        no server-side pagination.
      parameters:
        - name: origin
          in: query
          description: Filter by origin
          required: false
          schema:
            $ref: '#/components/schemas/ScenarioOrigin'
        - name: serviceType
          in: query
          description: Filter by service type
          required: false
          schema:
            $ref: '#/components/schemas/ServiceType'
        - name: peerId
          in: query
          description: Filter by bound peer
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Scenario list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ScenarioSummary'
        default:
          $ref: '#/components/responses/Problem'
    post:
      tags: [Scenarios]
      operationId: createScenario
      summary: Create a new scenario from scratch
      description: |
        Creates a user scenario. Most users will reach for
        `POST /scenarios/{id}/duplicate` instead to start from a system
        starter; this endpoint is for advanced flows that bring their own
        full scenario body.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScenarioInput'
      responses:
        '201':
          description: Scenario created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Scenario'
        '422':
          $ref: '#/components/responses/ValidationProblem'
        default:
          $ref: '#/components/responses/Problem'
  /scenarios/{id}:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    get:
      tags: [Scenarios]
      operationId: getScenario
      summary: Get a single scenario by id
      responses:
        '200':
          description: Scenario detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Scenario'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Problem'
    put:
      tags: [Scenarios]
      operationId: updateScenario
      summary: Replace a scenario
      description: |
        Full replacement. System starters (`origin: "system"`) are
        immutable and return `409 Conflict` on write — duplicate first.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScenarioInput'
      responses:
        '200':
          description: Scenario updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Scenario'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Scenario is a system starter and cannot be modified
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          $ref: '#/components/responses/ValidationProblem'
        default:
          $ref: '#/components/responses/Problem'
    delete:
      tags: [Scenarios]
      operationId: deleteScenario
      summary: Delete a scenario
      description: |
        User scenarios only. System starters cannot be deleted (returns
        `409 Conflict`).
      responses:
        '204':
          description: Scenario deleted
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Scenario is a system starter and cannot be deleted
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        default:
          $ref: '#/components/responses/Problem'
...
components:
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      description: Resource identifier
      schema:
        type: string
...
  responses:
    Problem:
      description: Error response (RFC 7807)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    ValidationProblem:
      description: |
        Validation error (RFC 7807). Carries field-level error messages
        keyed by JSON Pointer references into the request body.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  schemas:
...
    Problem:
      type: object
      description: RFC 7807 problem details
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          default: about:blank
        title:
          type: string
        status:
          type: integer
          minimum: 100
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri
        errors:
          type: object
          description: |
            Optional field-level validation errors. Keys are JSON Pointer
            references into the request body (e.g. `/name`, `/services/0/ratingGroup`);
            values are arrays of human-readable error messages. Present on
            `422` responses and ignorable otherwise.
          additionalProperties:
            type: array
            items:
              type: string
            minItems: 1

    # ─────────────── Pagination ───────────────
...
    SessionMode:
      type: string
      description: |
        Session-based vs event-based charging model.
        - `session` — INITIAL → UPDATE(s) → TERMINATE lifecycle.
        - `event`   — one-shot CCR-EVENT, no lifecycle.
      enum: [session, event]
    ServiceModel:
      type: string
      description: |
        Where service-unit AVPs live in the CCR.
        - `root`        — RSU/USU under the CCR root; no MSCC; MSI omitted.
        - `single-mscc` — exactly one MSCC block; MSI = 0.
        - `multi-mscc`  — one-to-many MSCC blocks; MSI = 1; OCTET only.
        See compatibility matrix in docs/ARCHITECTURE.md §4.
      enum: [root, single-mscc, multi-mscc]
    ScenarioOrigin:
      type: string
      description: |
        - `system` — ships with the product; immutable; can only be duplicated.
        - `user`   — created/owned by the user.
      enum: [system, user]
    ServiceType:
      type: string
      description: |
        Semantic service classification. Drives unit type, default session
        mode, and the Service-Information child AVP injected by the engine.
        Combined with `serviceProfile` for the AVP code mapping.

        | Value         | Unit type | Session mode | Service-Info child              |
        |---------------|-----------|--------------|----------------------------------|
        | VOICE         | TIME      | session      | IMS-Information / IN_INFORMATION |
        | DATA          | VOLUME    | session      | PS-Information (both profiles)   |
        | SMS           | EVENT     | session      | SMS-Information / SMS_INFORMATION|
        | USSD1_EVENT   | EVENT     | event        | USSD-Information / DCD_INFORMATION|
        | USSD1_SESSION | EVENT     | session      | USSD-Information / DCD_INFORMATION|
        | USSD2_SESSION | TIME      | session      | USSD-Information / DCD_INFORMATION|
      enum: [VOICE, DATA, SMS, USSD1_EVENT, USSD1_SESSION, USSD2_SESSION]
    ServiceProfile:
      type: string
      description: |
        OCS vendor profile. Selects the AVP code mapping for the
        `serviceType` → `Service-Information` child AVP.

        **3GPP** (vendor-id 10415):
        - VOICE         → IMS-Information (876)
        - DATA          → PS-Information (874)
        - SMS           → SMS-Information (2000)
        - USSD1_EVENT, USSD1_SESSION, USSD2_SESSION → USSD-Information (885)

        **HUAWEI** (Huawei-specific AVPs use vendor-id 2011):
        - VOICE         → IN_INFORMATION (20300)
        - DATA          → PS-Information (874, vendor 10415)
        - SMS           → SMS_INFORMATION (20400)
        - USSD1_EVENT, USSD1_SESSION, USSD2_SESSION → DCD_INFORMATION (2115)
      enum: [3GPP, HUAWEI]
      default: 3GPP
    RequestType:
      type: string
      enum: [INITIAL, UPDATE, TERMINATE, EVENT]
    VarRef:
      type: string
      description: |
        Variable reference. In `avpTree` leaf values the form `{{NAME}}`
        is used; in step configuration and expressions the bare `NAME`
        is used. Both forms resolve to the same variable.
      minLength: 1
...
    VariableSourceGenerator:
      type: object
      required: [kind, strategy, refresh]
      properties:
        kind:
          type: string
          const: generator
        strategy:
          $ref: '#/components/schemas/GeneratorStrategy'
        refresh:
          $ref: '#/components/schemas/GeneratorRefresh'
        params:
          type: object
          description: |
            Strategy-dependent parameters. Shape varies by `strategy`:
            - `literal`       — `{ value: <VarValue> }`
            - `incrementer`   — `{ start?: number, step?: number }`
            - `random-int`    — `{ min?: number, max?: number }`
            - `random-string` — `{ length: number, charset: "alpha"|"numeric"|"alphanumeric"|"hex" }`
            - `random-choice` — `{ options: string[] }`
            - `uuid`          — no params.
          additionalProperties: true
    VariableSourceBound:
      type: object
      required: [kind, from, field]
      properties:
        kind:
          type: string
          const: bound
        from:
          type: string
          enum: [subscriber, peer, config, step]
        field:
          type: string
          description: |
            Name of the field on the source resource, e.g. `msisdn` /
            `originHost` / `service-context` / `requestType`.
...
    VariableSource:
      description: |
        Discriminated union identifying where a variable's value comes from.
      oneOf:
        - $ref: '#/components/schemas/VariableSourceGenerator'
        - $ref: '#/components/schemas/VariableSourceBound'
        - $ref: '#/components/schemas/VariableSourceExtracted'
      discriminator:
        propertyName: kind
        mapping:
          generator: '#/components/schemas/VariableSourceGenerator'
          bound: '#/components/schemas/VariableSourceBound'
          extracted: '#/components/schemas/VariableSourceExtracted'
    Variable:
      type: object
      description: |
        A user-declared variable. System variables (SESSION_ID, MSISDN,
        RG<rg>_GRANTED, etc.) are auto-provisioned at run time and do not
        appear in `scenario.variables`.
      required: [name, source]
      properties:
        name:
          type: string
          description: |
            Unique within the scenario. Convention — `SCREAMING_SNAKE_CASE`.
            In `multi-mscc` mode, per-service variables are prefixed
            `RG<rg>_` (engine-provisioned — not authored here).
          pattern: '^[A-Z][A-Z0-9_]*$'
        source:
          $ref: '#/components/schemas/VariableSource'
        description:
          type: string

    # --- AVP tree --------------------------------------------------------
    AvpNode:
      type: object
      description: |
        One node in the CCR envelope. Either a grouped AVP (has `children`)
        or a leaf AVP (has `valueRef`). The `avpTree` is the CCR *minus*
        the service-unit AVPs — those are modelled separately in
        `scenario.services` and spliced in at send time by the engine
        (see docs/ARCHITECTURE.md §4 and §7).
      required: [name, code]
      properties:
        name:
          type: string
          description: AVP name from the Diameter dictionary (e.g. `Origin-Host`).
        code:
          type: integer
          description: AVP code (e.g. 264 for Origin-Host).
          minimum: 0
        vendorId:
          type: integer
          description: Vendor-Id, omitted for base-protocol AVPs.
          minimum: 0
        children:
          type: array
          description: |
            Grouped AVP body. Present iff this is a grouped AVP.
          items:
            $ref: '#/components/schemas/AvpNode'
        valueRef:
          $ref: '#/components/schemas/VarRef'
        locked:
          type: boolean
          description: |
            When true the Builder prevents deletion of this node and marks
            it with an [engine] badge. Leaf nodes with `locked: true` still
            allow their `valueRef` to be edited; locked grouped nodes are
            read-only (expand/collapse only). Used for system-mandatory AVPs
            (Origin-Host, Session-Id, Service-Information sub-tree, etc.).

    # --- Services --------------------------------------------------------
    Service:
      type: object
      description: |
        One service entry on a scenario. Controls how the engine
        constructs RSU/USU (and MSCC wrapping) at send time. See §4
        for how `serviceModel` picks between root / single-mscc /
        multi-mscc placement.
      required: [id, requestedUnits]
      properties:
        id:
          type: string
          description: |
            Stable key for this service. In `multi-mscc` mode, typically
            the Rating-Group as a decimal string (e.g. `"100"`) since it
            doubles as the catalogue key and feeds `RG<rg>_` variable
            naming.
        ratingGroup:
          $ref: '#/components/schemas/VarRef'
          description: |
            Rating-Group AVP (432). Required for `single-mscc` and
            `multi-mscc`; omitted for `root`.
        serviceIdentifier:
          $ref: '#/components/schemas/VarRef'
          description: |
            Service-Identifier AVP (439). Optional for MSCC-based service
            models; omitted for `root`.
        requestedUnits:
          $ref: '#/components/schemas/VarRef'
          description: |
            Variable carrying the requested-service-units quantity. Engine
            applies the §7 presence rules to decide whether RSU is emitted.
        usedUnits:
          $ref: '#/components/schemas/VarRef'
          description: |
            Variable carrying used-service-units quantity. Engine applies
            the §7 presence rules to decide whether USU is emitted.
...
    RequestStep:
      type: object
      required: [kind, requestType]
      properties:
        kind:
          type: string
          const: request
        label:
          type: string
          description: Optional display label shown in the execution history pane.
        requestType:
          $ref: '#/components/schemas/RequestType'
        services:
          $ref: '#/components/schemas/ServiceSelection'
          description: |
            Required in `multi-mscc` scenarios. Omitted in `root` and
            `single-mscc` (the single service is implicit).
        overrides:
          type: object
          description: |
            Per-step transient map of variable-name → value, shadowing
            scenario defaults for this step only. Keys must be declared
            variables.
          additionalProperties:
            $ref: '#/components/schemas/VarValue'
        assertions:
          type: array
          items:
            type: string
          description: |
            Post-receive predicates. On failure the step is marked failed
            and the first matching result handler decides what to do
            next (default: stop).
        guards:
          type: array
          items:
            type: string
          description: Pre-send predicates. On failure the step is skipped.
        resultHandlers:
          type: array
          items:
            $ref: '#/components/schemas/ResultHandler'
        repeat:
          type: integer
          minimum: 1
          description: |
            Fixed iteration count. Mutually exclusive with `repeatUntil`.
            Only meaningful for UPDATE steps.
        repeatUntil:
          type: string
          description: |
            Expression evaluated after each iteration. Loop exits when the
            expression returns true. Mutually exclusive with `repeat`.
            Both `{{VAR}}` and bare `VAR` notation are accepted.
        maxRepeat:
          type: integer
          minimum: 1
          description: |
            Safety cap when `repeatUntil` is set. Loop exits after this many
            iterations even if the condition hasn't fired. Defaults to 100.
        delaySec:
          type: integer
          minimum: 0
          description: |
            Fixed inter-iteration delay in seconds. Applied between
            consecutive iterations of a repeat loop.
        delayJitterSec:
          type: integer
          minimum: 0
          description: |
            Random jitter added to `delaySec`. Actual delay is
            `delaySec + rand(0, delayJitterSec)` seconds.
        useValidityTime:
          type: boolean
          description: |
            When true, the inter-iteration delay is derived from the
            Validity-Time AVP in the last CCA response, scaled by
            `validityScale`. Overrides `delaySec`/`delayJitterSec`.
        validityScale:
          type: number
          description: |
            Multiplier applied to the Validity-Time value when
            `useValidityTime` is true. Defaults to 1.0.
    ScenarioStep:
      description: A scenario step. All steps are request steps.
      $ref: '#/components/schemas/RequestStep'

    # --- Scenario top-level ----------------------------------------------
    ScenarioSummary:
      type: object
      description: Row-level shape for the scenarios list view.
      required: [id, name, sessionMode, serviceModel, origin, stepCount, updatedAt]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        serviceType:
          $ref: '#/components/schemas/ServiceType'
        serviceProfile:
          $ref: '#/components/schemas/ServiceProfile'
        sessionMode:
          $ref: '#/components/schemas/SessionMode'
        serviceModel:
          $ref: '#/components/schemas/ServiceModel'
        origin:
          $ref: '#/components/schemas/ScenarioOrigin'
        favourite:
          type: boolean
        subscriberId:
          type: string
        peerId:
          type: string
        stepCount:
          type: integer
          minimum: 0
        updatedAt:
          type: string
          format: date-time
    Scenario:
      description: Full scenario body.
      allOf:
        - $ref: '#/components/schemas/ScenarioSummary'
        - type: object
          required: [avpTree, services, variables, steps]
          properties:
            avpTree:
              type: array
              description: |
                Root-level AVPs of the CCR envelope, minus the service-unit
                AVPs. Structure is frozen at execution start.
              items:
                $ref: '#/components/schemas/AvpNode'
            services:
              type: array
              description: |
                Service entries. Exactly one in `root` / `single-mscc`;
                one or more in `multi-mscc`.
              items:
                $ref: '#/components/schemas/Service'
              minItems: 1
            variables:
              type: array
              description: |
                User-declared variables. System variables are implicit
                and auto-provisioned at run time.
              items:
                $ref: '#/components/schemas/Variable'
            steps:
              type: array
              items:
                $ref: '#/components/schemas/ScenarioStep'
    ScenarioInput:
      type: object
      description: |
        Writable subset of `Scenario`. Used for create (POST /scenarios)
        and replace (PUT /scenarios/{id}). Server-owned fields (`id`,
        `origin`, `updatedAt`, `stepCount`) are ignored on write —
        `origin` is always `user` on create/replace; system starters
        cannot be updated.
      required:
        [name, sessionMode, serviceModel, subscriberId, peerId, avpTree, services, variables, steps]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 128
        description:
          type: string
        serviceType:
          $ref: '#/components/schemas/ServiceType'
        serviceProfile:
          $ref: '#/components/schemas/ServiceProfile'
        sessionMode:
          $ref: '#/components/schemas/SessionMode'
        serviceModel:
          $ref: '#/components/schemas/ServiceModel'
        favourite:
          type: boolean
          default: false
        subscriberId:
          type: string
        peerId:
          type: string
        avpTree:
          type: array
          items:
            $ref: '#/components/schemas/AvpNode'
        services:
          type: array
          items:
            $ref: '#/components/schemas/Service'
          minItems: 1
        variables:
          type: array
          items:
            $ref: '#/components/schemas/Variable'
        steps:
          type: array
          items:
            $ref: '#/components/schemas/ScenarioStep'
...